3691014: Prevent unsafe narrowing: ui/accessibility, part 2
Ref: https://chromium-review.googlesource.com/c/chromium/src/+/3691014
(cherry picked from commit fca99a5cf992dfe0aec7ff832ee7775c7ddf1322)
(cherry picked from commit 08f7a35765fde0467ef44d773e943cf5998619ae)

+++ shell/browser/ui/webui/accessibility_ui.cc
@@ -196,7 +196,7 @@ std::string RecursiveDumpAXPlatformNodeAsString(
     }
   }
   str += "\n";
-  for (int i = 0; i < node->GetDelegate()->GetChildCount(); i++) {
+  for (size_t i = 0; i < node->GetDelegate()->GetChildCount(); i++) {
     gfx::NativeViewAccessible child = node->GetDelegate()->ChildAtIndex(i);
     const ui::AXPlatformNode* child_node =
         ui::AXPlatformNode::FromNativeViewAccessible(child);
